refactor: decompose post component into multiple files
